converted the menus module to use the action controllers
