+++ ACP3/Modules/ACP3/Errors/Controller/Frontend/Index/Index.php
@@ -1,12 +1,15 @@
 <?php
+/**
+ * Copyright (c) 2016 by the ACP3 Developers. See the LICENCE file at the top-level module directory for licencing details.
+ */
 
-namespace ACP3\Modules\ACP3\Errors\Controller;
+namespace ACP3\Modules\ACP3\Errors\Controller\Frontend\Index;
 
 use ACP3\Core;
 
 /**
  * Class Index
- * @package ACP3\Modules\ACP3\Errors\Controller
+ * @package ACP3\Modules\ACP3\Errors\Controller\Frontend\Index
  */
 class Index extends Core\Modules\FrontendController
 {
