+++ ACP3/Modules/ACP3/Errors/Resources/config/services.yml
@@ -1,5 +1,5 @@
 parameters:
-  errors.controller.frontend.index.fqdn: 'ACP3\Modules\ACP3\Errors\Controller\Index'
+  errors.controller.frontend.index.fqdn: 'ACP3\Modules\ACP3\Errors\Controller\Frontend\Index\Index'
   errors.installer.schema.fqdn: 'ACP3\Modules\ACP3\Errors\Installer\Schema'
   errors.installer.migration.fqdn: 'ACP3\Modules\ACP3\Errors\Installer\Migration'
