--- ACP3/Modules/ACP3/Menus/Controller/Admin/Index.php
@@ -1,242 +1,0 @@
-<?php
-
-namespace ACP3\Modules\ACP3\Menus\Controller\Admin;
-
-use ACP3\Core;
-use ACP3\Modules\ACP3\Menus;
-
-/**
- * Class Index
- * @package ACP3\Modules\ACP3\Menus\Controller\Admin
- */
-class Index extends Core\Modules\AdminController
-{
-    /**
-     * @var \ACP3\Core\Helpers\FormToken
-     */
-    protected $formTokenHelper;
-    /**
-     * @var \ACP3\Core\NestedSet
-     */
-    protected $nestedSet;
-    /**
-     * @var \ACP3\Modules\ACP3\Menus\Helpers\MenuItemsList
-     */
-    protected $menusHelpers;
-    /**
-     * @var \ACP3\Modules\ACP3\Menus\Model\MenuRepository
-     */
-    protected $menuRepository;
-    /**
-     * @var \ACP3\Modules\ACP3\Menus\Cache
-     */
-    protected $menusCache;
-    /**
-     * @var \ACP3\Modules\ACP3\Menus\Validation\MenuFormValidation
-     */
-    protected $menuFormValidation;
-    /**
-     * @var \ACP3\Modules\ACP3\Menus\Model\MenuItemRepository
-     */
-    protected $menuItemRepository;
-
-    /**
-     * @param \ACP3\Core\Modules\Controller\AdminContext             $context
-     * @param \ACP3\Core\NestedSet                                   $nestedSet
-     * @param \ACP3\Core\Helpers\FormToken                           $formTokenHelper
-     * @param \ACP3\Modules\ACP3\Menus\Helpers\MenuItemsList         $menusHelpers
-     * @param \ACP3\Modules\ACP3\Menus\Model\MenuRepository          $menuRepository
-     * @param \ACP3\Modules\ACP3\Menus\Model\MenuItemRepository      $menuItemRepository
-     * @param \ACP3\Modules\ACP3\Menus\Cache                         $menusCache
-     * @param \ACP3\Modules\ACP3\Menus\Validation\MenuFormValidation $menuFormValidation
-     */
-    public function __construct(
-        Core\Modules\Controller\AdminContext $context,
-        Core\NestedSet $nestedSet,
-        Core\Helpers\FormToken $formTokenHelper,
-        Menus\Helpers\MenuItemsList $menusHelpers,
-        Menus\Model\MenuRepository $menuRepository,
-        Menus\Model\MenuItemRepository $menuItemRepository,
-        Menus\Cache $menusCache,
-        Menus\Validation\MenuFormValidation $menuFormValidation
-    ) {
-        parent::__construct($context);
-
-        $this->nestedSet = $nestedSet;
-        $this->formTokenHelper = $formTokenHelper;
-        $this->menusHelpers = $menusHelpers;
-        $this->menuRepository = $menuRepository;
-        $this->menuItemRepository = $menuItemRepository;
-        $this->menusCache = $menusCache;
-        $this->menuFormValidation = $menuFormValidation;
-    }
-
-    /**
-     * @return array|\Symfony\Component\HttpFoundation\RedirectResponse
-     */
-    public function actionCreate()
-    {
-        if ($this->request->getPost()->isEmpty() === false) {
-            return $this->_createPost($this->request->getPost()->all());
-        }
-
-        $this->formTokenHelper->generateFormToken();
-
-        return [
-            'form' => array_merge(['index_name' => '', 'title' => ''], $this->request->getPost()->all())
-        ];
-    }
-
-    /**
-     * @param string $action
-     *
-     * @return mixed
-     */
-    public function actionDelete($action = '')
-    {
-        return $this->actionHelper->handleDeleteAction(
-            $this,
-            $action,
-            function ($items) {
-                $bool = false;
-
-                foreach ($items as $item) {
-                    if (!empty($item) && $this->menuRepository->menuExists($item) === true) {
-                        // Delete the assigned menu items and update the nested set tree
-                        $items = $this->menuItemRepository->getAllItemsByBlockId($item);
-                        foreach ($items as $row) {
-                            $this->nestedSet->deleteNode(
-                                $row['id'],
-                                Menus\Model\MenuItemRepository::TABLE_NAME,
-                                true
-                            );
-                        }
-
-                        $block = $this->menuRepository->getMenuNameById($item);
-                        $bool = $this->menuRepository->delete($item);
-                        $this->menusCache->getCacheDriver()->delete(Menus\Cache::CACHE_ID_VISIBLE . $block);
-                    }
-                }
-
-                $this->menusCache->saveMenusCache();
-
-                return $bool;
-            }
-        );
-    }
-
-    /**
-     * @param int $id
-     *
-     * @return array|\Symfony\Component\HttpFoundation\RedirectResponse
-     * @throws \ACP3\Core\Exceptions\ResultNotExists
-     */
-    public function actionEdit($id)
-    {
-        $menu = $this->menuRepository->getOneById($id);
-
-        if (empty($menu) === false) {
-            $this->breadcrumb->setTitlePostfix($menu['title']);
-
-            if ($this->request->getPost()->isEmpty() === false) {
-                return $this->_editPost($this->request->getPost()->all(), $id);
-            }
-
-            $this->formTokenHelper->generateFormToken();
-
-            return [
-                'form' => array_merge($menu, $this->request->getPost()->all())
-            ];
-        }
-
-        throw new Core\Exceptions\ResultNotExists();
-    }
-
-    public function actionIndex()
-    {
-        $menus = $this->menuRepository->getAllMenus();
-        $c_menus = count($menus);
-
-        if ($c_menus > 0) {
-            $canDeleteItem = $this->acl->hasPermission('admin/menus/items/delete');
-            $canEditItem = $this->acl->hasPermission('admin/menus/items/edit');
-            $canSortItem = $this->acl->hasPermission('admin/menus/items/order');
-            $this->view->assign('can_delete_item', $canDeleteItem);
-            $this->view->assign('can_edit_item', $canEditItem);
-            $this->view->assign('can_order_item', $canSortItem);
-            $this->view->assign('can_delete', $this->acl->hasPermission('admin/menus/index/delete'));
-            $this->view->assign('can_edit', $this->acl->hasPermission('admin/menus/index/edit'));
-
-            $colspan = 4;
-            if ($canDeleteItem || $canEditItem) {
-                $colspan += 1;
-            }
-            if ($canSortItem) {
-                $colspan += 1;
-            }
-            $this->view->assign('colspan', $colspan);
-
-            $menuItems = $this->menusHelpers->menuItemsList();
-            for ($i = 0; $i < $c_menus; ++$i) {
-                if (isset($menuItems[$menus[$i]['index_name']]) === false) {
-                    $menuItems[$menus[$i]['index_name']]['title'] = $menus[$i]['title'];
-                    $menuItems[$menus[$i]['index_name']]['menu_id'] = $menus[$i]['id'];
-                    $menuItems[$menus[$i]['index_name']]['items'] = [];
-                }
-            }
-            $this->view->assign('pages_list', $menuItems);
-        }
-    }
-
-    /**
-     * @param array $formData
-     *
-     * @return \Symfony\Component\HttpFoundation\RedirectResponse
-     */
-    protected function _createPost(array $formData)
-    {
-        return $this->actionHelper->handleCreatePostAction(function () use ($formData) {
-            $this->menuFormValidation->validate($formData);
-
-            $insertValues = [
-                'id' => '',
-                'index_name' => $formData['index_name'],
-                'title' => Core\Functions::strEncode($formData['title']),
-            ];
-
-            $lastId = $this->menuRepository->insert($insertValues);
-
-            $this->formTokenHelper->unsetFormToken();
-
-            return $lastId;
-        });
-    }
-
-    /**
-     * @param array $formData
-     * @param int   $id
-     *
-     * @return \Symfony\Component\HttpFoundation\RedirectResponse
-     */
-    protected function _editPost(array $formData, $id)
-    {
-        return $this->actionHelper->handleEditPostAction(function () use ($formData, $id) {
-            $this->menuFormValidation
-                ->setMenuId($id)
-                ->validate($formData);
-
-            $updateValues = [
-                'index_name' => $formData['index_name'],
-                'title' => Core\Functions::strEncode($formData['title']),
-            ];
-
-            $bool = $this->menuRepository->update($updateValues, $id);
-
-            $this->menusCache->saveMenusCache();
-
-            $this->formTokenHelper->unsetFormToken();
-
-            return $bool;
-        });
-    }
-}
