+++ ACP3/Modules/ACP3/Menus/Controller/Admin/Index/Create.php
@@ -1,0 +1,90 @@
+<?php
+/**
+ * Copyright (c) 2016 by the ACP3 Developers. See the LICENCE file at the top-level module directory for licencing details.
+ */
+
+namespace ACP3\Modules\ACP3\Menus\Controller\Admin\Index;
+
+use ACP3\Core;
+use ACP3\Modules\ACP3\Menus;
+
+/**
+ * Class Create
+ * @package ACP3\Modules\ACP3\Menus\Controller\Admin\Index
+ */
+class Create extends Core\Modules\AdminController
+{
+    /**
+     * @var \ACP3\Core\Helpers\FormToken
+     */
+    protected $formTokenHelper;
+    /**
+     * @var \ACP3\Modules\ACP3\Menus\Model\MenuRepository
+     */
+    protected $menuRepository;
+    /**
+     * @var \ACP3\Modules\ACP3\Menus\Validation\MenuFormValidation
+     */
+    protected $menuFormValidation;
+
+    /**
+     * Create constructor.
+     *
+     * @param \ACP3\Core\Modules\Controller\AdminContext             $context
+     * @param \ACP3\Core\Helpers\FormToken                           $formTokenHelper
+     * @param \ACP3\Modules\ACP3\Menus\Model\MenuRepository          $menuRepository
+     * @param \ACP3\Modules\ACP3\Menus\Validation\MenuFormValidation $menuFormValidation
+     */
+    public function __construct(
+        Core\Modules\Controller\AdminContext $context,
+        Core\Helpers\FormToken $formTokenHelper,
+        Menus\Model\MenuRepository $menuRepository,
+        Menus\Validation\MenuFormValidation $menuFormValidation
+    ) {
+        parent::__construct($context);
+
+        $this->formTokenHelper = $formTokenHelper;
+        $this->menuRepository = $menuRepository;
+        $this->menuFormValidation = $menuFormValidation;
+    }
+
+    /**
+     * @return array|\Symfony\Component\HttpFoundation\RedirectResponse
+     */
+    public function execute()
+    {
+        if ($this->request->getPost()->isEmpty() === false) {
+            return $this->executePost($this->request->getPost()->all());
+        }
+
+        $this->formTokenHelper->generateFormToken();
+
+        return [
+            'form' => array_merge(['index_name' => '', 'title' => ''], $this->request->getPost()->all())
+        ];
+    }
+
+    /**
+     * @param array $formData
+     *
+     * @return \Symfony\Component\HttpFoundation\RedirectResponse
+     */
+    protected function executePost(array $formData)
+    {
+        return $this->actionHelper->handleCreatePostAction(function () use ($formData) {
+            $this->menuFormValidation->validate($formData);
+
+            $insertValues = [
+                'id' => '',
+                'index_name' => $formData['index_name'],
+                'title' => Core\Functions::strEncode($formData['title']),
+            ];
+
+            $lastId = $this->menuRepository->insert($insertValues);
+
+            $this->formTokenHelper->unsetFormToken();
+
+            return $lastId;
+        });
+    }
+}
