+++ ACP3/Modules/ACP3/Menus/Controller/Admin/Index/Edit.php
@@ -1,0 +1,112 @@
+<?php
+/**
+ * Copyright (c) 2016 by the ACP3 Developers. See the LICENCE file at the top-level module directory for licencing details.
+ */
+
+namespace ACP3\Modules\ACP3\Menus\Controller\Admin\Index;
+
+use ACP3\Core;
+use ACP3\Modules\ACP3\Menus;
+
+/**
+ * Class Edit
+ * @package ACP3\Modules\ACP3\Menus\Controller\Admin\Index
+ */
+class Edit extends Core\Modules\AdminController
+{
+    /**
+     * @var \ACP3\Core\Helpers\FormToken
+     */
+    protected $formTokenHelper;
+    /**
+     * @var \ACP3\Modules\ACP3\Menus\Model\MenuRepository
+     */
+    protected $menuRepository;
+    /**
+     * @var \ACP3\Modules\ACP3\Menus\Cache
+     */
+    protected $menusCache;
+    /**
+     * @var \ACP3\Modules\ACP3\Menus\Validation\MenuFormValidation
+     */
+    protected $menuFormValidation;
+
+    /**
+     * Edit constructor.
+     *
+     * @param \ACP3\Core\Modules\Controller\AdminContext             $context
+     * @param \ACP3\Core\Helpers\FormToken                           $formTokenHelper
+     * @param \ACP3\Modules\ACP3\Menus\Model\MenuRepository          $menuRepository
+     * @param \ACP3\Modules\ACP3\Menus\Cache                         $menusCache
+     * @param \ACP3\Modules\ACP3\Menus\Validation\MenuFormValidation $menuFormValidation
+     */
+    public function __construct(
+        Core\Modules\Controller\AdminContext $context,
+        Core\Helpers\FormToken $formTokenHelper,
+        Menus\Model\MenuRepository $menuRepository,
+        Menus\Cache $menusCache,
+        Menus\Validation\MenuFormValidation $menuFormValidation
+    ) {
+        parent::__construct($context);
+
+        $this->formTokenHelper = $formTokenHelper;
+        $this->menuRepository = $menuRepository;
+        $this->menusCache = $menusCache;
+        $this->menuFormValidation = $menuFormValidation;
+    }
+
+    /**
+     * @param int $id
+     *
+     * @return array|\Symfony\Component\HttpFoundation\RedirectResponse
+     * @throws \ACP3\Core\Exceptions\ResultNotExists
+     */
+    public function execute($id)
+    {
+        $menu = $this->menuRepository->getOneById($id);
+
+        if (empty($menu) === false) {
+            $this->breadcrumb->setTitlePostfix($menu['title']);
+
+            if ($this->request->getPost()->isEmpty() === false) {
+                return $this->executePost($this->request->getPost()->all(), $id);
+            }
+
+            $this->formTokenHelper->generateFormToken();
+
+            return [
+                'form' => array_merge($menu, $this->request->getPost()->all())
+            ];
+        }
+
+        throw new Core\Exceptions\ResultNotExists();
+    }
+
+    /**
+     * @param array $formData
+     * @param int   $id
+     *
+     * @return \Symfony\Component\HttpFoundation\RedirectResponse
+     */
+    protected function executePost(array $formData, $id)
+    {
+        return $this->actionHelper->handleEditPostAction(function () use ($formData, $id) {
+            $this->menuFormValidation
+                ->setMenuId($id)
+                ->validate($formData);
+
+            $updateValues = [
+                'index_name' => $formData['index_name'],
+                'title' => Core\Functions::strEncode($formData['title']),
+            ];
+
+            $bool = $this->menuRepository->update($updateValues, $id);
+
+            $this->menusCache->saveMenusCache();
+
+            $this->formTokenHelper->unsetFormToken();
+
+            return $bool;
+        });
+    }
+}
