+++ ACP3/Modules/ACP3/Menus/Resources/config/controllers.yml
@@ -1,0 +1,88 @@
+parameters:
+  menus.controller.admin.index.create.fqdn: 'ACP3\Modules\ACP3\Menus\Controller\Admin\Index\Create'
+  menus.controller.admin.index.delete.fqdn: 'ACP3\Modules\ACP3\Menus\Controller\Admin\Index\Delete'
+  menus.controller.admin.index.edit.fqdn: 'ACP3\Modules\ACP3\Menus\Controller\Admin\Index\Edit'
+  menus.controller.admin.index.index.fqdn: 'ACP3\Modules\ACP3\Menus\Controller\Admin\Index\Index'
+  menus.controller.admin.items.create.fqdn: 'ACP3\Modules\ACP3\Menus\Controller\Admin\Items\Create'
+  menus.controller.admin.items.delete.fqdn: 'ACP3\Modules\ACP3\Menus\Controller\Admin\Items\Delete'
+  menus.controller.admin.items.edit.fqdn: 'ACP3\Modules\ACP3\Menus\Controller\Admin\Items\Edit'
+  menus.controller.admin.items.order.fqdn: 'ACP3\Modules\ACP3\Menus\Controller\Admin\Items\Order'
+
+services:
+  menus.controller.admin.index.create:
+    class: %menus.controller.admin.index.create.fqdn%
+    arguments:
+      - @core.context.admin
+      - @core.helpers.formToken
+      - @menus.model.menurepository
+      - @menus.validator.menu
+
+  menus.controller.admin.index.delete:
+    class: %menus.controller.admin.index.delete.fqdn%
+    arguments:
+      - @core.context.admin
+      - @core.nested_set
+      - @menus.model.menurepository
+      - @menus.model.menuitemrepository
+      - @menus.cache
+
+  menus.controller.admin.index.edit:
+    class: %menus.controller.admin.index.edit.fqdn%
+    arguments:
+      - @core.context.admin
+      - @core.helpers.formToken
+      - @menus.model.menurepository
+      - @menus.cache
+      - @menus.validator.menu
+
+  menus.controller.admin.index.index:
+    class: %menus.controller.admin.index.index.fqdn%
+    arguments:
+      - @core.context.admin
+      - @menus.helpers.menuitemslist
+      - @menus.model.menurepository
+
+  menus.controller.admin.items.create:
+    class: %menus.controller.admin.items.create.fqdn%
+    arguments:
+      - @core.context.admin
+      - @core.router.aliases
+      - @core.nested_set
+      - @core.helpers.formToken
+      - @menus.model.menurepository
+      - @menus.model.menuitemrepository
+      - @menus.cache
+      - @menus.helpers.menuitemformfields
+      - @menus.validator.menuitem
+    calls:
+      - ['setArticlesHelpers', ['@?articles.helpers']]
+
+  menus.controller.admin.items.delete:
+    class: %menus.controller.admin.items.delete.fqdn%
+    arguments:
+      - @core.context.admin
+      - @core.nested_set
+      - @menus.model.menuitemrepository
+      - @menus.cache
+
+  menus.controller.admin.items.edit:
+    class: %menus.controller.admin.items.edit.fqdn%
+    arguments:
+      - @core.context.admin
+      - @core.router.aliases
+      - @core.nested_set
+      - @core.helpers.formToken
+      - @menus.model.menuitemrepository
+      - @menus.cache
+      - @menus.helpers.menuitemformfields
+      - @menus.validator.menuitem
+    calls:
+      - ['setArticlesHelpers', ['@?articles.helpers']]
+
+  menus.controller.admin.items.order:
+    class: %menus.controller.admin.items.order.fqdn%
+    arguments:
+      - @core.context.admin
+      - @core.nested_set
+      - @menus.model.menuitemrepository
+      - @menus.cache
